feat(instantsearch): remove Instance Service
