+++ src/instantsearch/instantsearch.ts
@@ -5,14 +5,61 @@ import {
   OnInit,
   OnDestroy,
   Output,
-  EventEmitter
+  EventEmitter,
+  Inject,
+  PLATFORM_ID
 } from "@angular/core";
+import { isPlatformBrowser } from "@angular/common";
 
-import {
-  NgAisInstance,
-  InstantSearchConfig,
-  InstantSearchInstance
-} from "./instantsearch-instance";
+import instantsearch from "instantsearch.js/es";
+
+import { Widget } from "../base-widget";
+import { VERSION } from "../version";
+
+export type InstantSearchConfig = {
+  appId: string;
+  apiKey: string;
+  indexName: string;
+
+  numberLocale?: string;
+  searchFunction?: () => void;
+  createAlgoliaClient?: (
+    algoliasearch: Function,
+    appId: string,
+    apiKey: string
+  ) => object;
+  searchParameters?: object | void;
+  urlSync?:
+    | boolean
+    | {
+        mapping?: object;
+        threshold?: number;
+        trackedParameters?: string[];
+        useHash?: boolean;
+        getHistoryState?: () => object;
+      };
+};
+
+export class InstantSearchInstance {
+  public start: () => void;
+
+  public addWidget: (widget: Widget) => void;
+  public addWidgets: (widgets: Widget[]) => void;
+
+  public removeWidget: (widget: Widget) => void;
+  public removeWidgets: (widgets: Widget[]) => void;
+
+  // EventEmmiter
+  public on: (eventName: string, callback: Function) => void;
+  public removeListener: (eventName: string, callback: Function) => void;
+
+  public helper: {
+    lastResults: Object;
+    state: Object;
+  };
+
+  public dispose: () => void;
+}
 
 @Component({
   selector: "ng-ais-instantsearch",
@@ -28,32 +75,61 @@ export class NgAisInstantSearch implements AfterViewInit, OnInit, OnDestroy {
     state: {};
   }>();
 
-  public searchInstance: InstantSearchInstance;
+  public instantSearchInstance: InstantSearchInstance;
 
-  constructor(private instantSearchInstances: NgAisInstance) {}
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
 
   public ngOnInit() {
-    this.searchInstance = this.instantSearchInstances.init(
-      this.config,
-      this.instanceName
-    );
-
-    this.searchInstance.on("render", this.onInstantSearchRender);
+    this.createInstantSearchInstance(this.config);
   }
 
   public ngAfterViewInit() {
-    this.searchInstance.start();
+    this.instantSearchInstance.start();
   }
 
   public ngOnDestroy() {
-    this.searchInstance.removeListener("render", this.onInstantSearchRender);
-    this.instantSearchInstances.dispose(this.instanceName);
+    this.instantSearchInstance.removeListener("render", this.onRender);
+    this.instantSearchInstance.dispose();
   }
 
-  onInstantSearchRender = () => {
+  public createInstantSearchInstance(config: InstantSearchConfig) {
+    // add default searchParameters with highlighting config
+    if (!config.searchParameters) config.searchParameters = {};
+    Object.assign(config.searchParameters, {
+      highlightPreTag: "__ais-highlight__",
+      highlightPostTag: "__/ais-highlight__"
+    });
+
+    // remove URLSync widget if on SSR
+    if (!isPlatformBrowser(this.platformId)) {
+      config.urlSync = false;
+    }
+
+    // custom algolia client agent
+    if (!config.createAlgoliaClient) {
+      config.createAlgoliaClient = (algoliasearch, appId, apiKey) => {
+        const client = algoliasearch(appId, apiKey);
+        client.addAlgoliaAgent(`angular-instantsearch ${VERSION}`);
+        return client;
+      };
+    }
+
+    this.instantSearchInstance = instantsearch(config);
+    this.instantSearchInstance.on("render", this.onRender);
+  }
+
+  public addWidget(widget: Widget) {
+    this.instantSearchInstance.addWidget(widget);
+  }
+
+  public removeWidget(widget: Widget) {
+    this.instantSearchInstance.removeWidget(widget);
+  }
+
+  onRender = () => {
     this.change.emit({
-      results: this.searchInstance.helper.lastResults,
-      state: this.searchInstance.helper.state
+      results: this.instantSearchInstance.helper.lastResults,
+      state: this.instantSearchInstance.helper.state
     });
   };
 }
